revert: ⏪ revert changes to ludus host firewall
if you are giving users SSH access to the ludus host, you can manually firewall them off. Due to the fact ludus can be installed on existing proxmox hosts, automatically changing the host firewall could have unintended consequences (in testing it caused lots of issues when the clients were on a 10.x network, although they were fixed). Documentation will be provided on how to firewall the host, but it will not be automated.

re #26

Signed-off-by: Erik [Bad Sector Labs] <555113-badsectorlabs@users.noreply.gitlab.com>